+++ app/layout.tsx
@@ -1,0 +1,25 @@
+import Footer from "../components/Footer"
+import NavBar from "../components/navbar"
+import './globals.css'
+
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <NavBar/>
+        {children}
+        <Footer />
+      </body>
+
+    </html>
+  )
+}
