+++ components/animated-element.tsx
