--- src/app/layout.tsx
@@ -1,25 +1,0 @@
-import type { Metadata } from "next";
-import "./globals.css";
-import NavBar from "@/components/navbar";
-import Footer from "@/components/footer";
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body >
-        <NavBar />
-        {children}
-        <Footer />
-      </body>
-    </html>
-  );
-}
